DRY similar code in all payment methods
Extract code into new superclass
