+++ spec/models/spree/payment_method/sisow_billing/sofort_spec.rb
@@ -1,9 +1,7 @@
 require 'spec_helper'
 
 describe Spree::PaymentMethod::SisowBilling::Sofort, type: :model do
-  let(:subject) { Spree::PaymentMethod::SisowBilling::Sofort.new }
   let(:order) { double("Spree::Order") }
-  let(:sisow_transaction) { double("Spree::SisowTransaction") }
   let(:payment){ double("Spree::Payment") }
 
   let(:options) {
@@ -30,23 +28,5 @@ describe Spree::PaymentMethod::SisowBilling::Sofort, type: :model do
     #payment.should_receive(:pend!)
 
     expect(subject.redirect_url(order, options)).to match(/https:\/\/www\.sisow\.nl\/Sisow\/iDeal\/Simulator\.aspx/)
-  end
-
-  it "should respond with false when calling payment_profiles_supported?" do
-    expect(subject.payment_profiles_supported?).to be false
-  end
-
-  it "should respond with true when calling auto_capture?" do
-    expect(subject.auto_capture?).to be true
-  end
-
-  it "should respond with true when the transaction was successfull" do
-    allow(sisow_transaction).to receive_message_chain(:status, :downcase).and_return('success')
-    expect(subject.purchase('123', sisow_transaction, {}).success?).to be true
-  end
-
-  it "should respond with false when the transaction was unsuccessfull" do
-    allow(sisow_transaction).to receive_message_chain(:status, :downcase).and_return('expired')
-    expect(subject.purchase('123', sisow_transaction, {}).success?).to be false
   end
 end
